Put the DynamoDB logic in a common file
