+++ lambdas/update_dynamo_capacity/dynamo_utils.py
@@ -1,0 +1,1 @@
+../common/dynamo_utils.py
